mini-project/requirements.txt
fastapi
uvicorn
spacy
scikit-learn
pandas
numpy
joblib
psycopg2-binary